File structure change to vendor related directory
