+++ sdks/python/apache_beam/tests/pipeline_verifiers.py
@@ -22,15 +22,13 @@ of test pipeline job. Customized verifier should extend
 `hamcrest.core.base_matcher.BaseMatcher` and override _matches.
 """
 
-import hashlib
 import logging
 
-from google.cloud import bigquery
-from google.cloud.exceptions import GoogleCloudError
 from hamcrest.core.base_matcher import BaseMatcher
 
 from apache_beam.io.fileio import ChannelFactory
 from apache_beam.runners.runner import PipelineState
+from apache_beam.tests import test_utils as utils
 from apache_beam.utils import retry
 
 try:
@@ -38,7 +36,6 @@ try:
 except ImportError:
   HttpError = None
 
-DEFAULT_HASHING_ALG = 'sha1'
 MAX_RETRIES = 4
 
 
@@ -75,24 +72,6 @@ def retry_on_io_error_and_server_error(exception):
     return False
 
 
-def retry_on_http_and_value_error(exception):
-  """Filter allowing retries on Bigquery errors and value error."""
-  if isinstance(exception, GoogleCloudError) or \
-          isinstance(exception, ValueError):
-    return True
-  else:
-    return False
-
-
-def compute_hash(list, hashing_alg=DEFAULT_HASHING_ALG):
-  """Compute a hash value from a list of string using SHA-1 as default."""
-  list.sort()
-  m = hashlib.new(hashing_alg)
-  for elem in list:
-    m.update(str(elem))
-  return m.hexdigest()
-
-
 class FileChecksumMatcher(BaseMatcher):
   """Matcher that verifies file(s) content by comparing file checksum.
 
@@ -124,72 +103,10 @@ class FileChecksumMatcher(BaseMatcher):
     read_lines = self._read_with_retry()
 
     # Compute checksum
-    self.checksum = compute_hash(read_lines)
+    self.checksum = utils.compute_hash(read_lines)
     logging.info('Read from given path %s, %d lines, checksum: %s.',
                  self.file_path, len(read_lines), self.checksum)
     return self.checksum == self.expected_checksum
-
-  def describe_to(self, description):
-    description \
-      .append_text("Expected checksum is ") \
-      .append_text(self.expected_checksum)
-
-  def describe_mismatch(self, pipeline_result, mismatch_description):
-    mismatch_description \
-      .append_text("Actual checksum is ") \
-      .append_text(self.checksum)
-
-
-class BigqueryMatcher(BaseMatcher):
-  """Matcher that verifies Bigquery data with given query.
-
-  Fetch Bigquery data with given query, compute a SHA-1 hash string and
-  compare with expected checksum.
-  """
-
-  def __init__(self, project, query, checksum):
-    if not query or not isinstance(query, str):
-      raise ValueError(
-          'Invalid argument: query. Please use non-empty string')
-    if not checksum or not isinstance(checksum, str):
-      raise ValueError(
-          'Invalid argument: checksum. Please use non-empty string')
-    self.project = project
-    self.query = query
-    self.expected_checksum = checksum
-
-  def _matches(self, _):
-    logging.info('Start verify Bigquery data.')
-    # Run query
-    bigquery_client = bigquery.Client(project=self.project)
-    response = self._query_with_retry(bigquery_client)
-
-    # Compute checksum
-    self.checksum = compute_hash(response)
-    logging.info('Read from given query (%s), total rows %d, checksum %s',
-                 self.query, len(response), self.checksum)
-
-    # Verify result
-    return self.checksum == self.expected_checksum
-
-  @retry.with_exponential_backoff(
-      num_retries=MAX_RETRIES,
-      retry_filter=retry_on_http_and_value_error)
-  def _query_with_retry(self, bigquery_client):
-    """Run Bigquery query with retry if got error http response"""
-    query = bigquery_client.run_sync_query(self.query)
-    query.run()
-
-    # Fetch query data one page at a time.
-    page_token = None
-    results = []
-    while True:
-      rows, _, page_token = query.fetch_data(page_token=page_token)
-      results.extend(rows)
-      if not page_token:
-        break
-
-    return results
 
   def describe_to(self, description):
     description \
