+++ sdks/python/apache_beam/tests/pipeline_verifiers_test.py
@@ -21,8 +21,6 @@ import logging
 import tempfile
 import unittest
 
-from google.cloud import bigquery
-from google.cloud.exceptions import NotFound
 from hamcrest import assert_that as hc_assert_that
 from mock import Mock, patch
 
@@ -119,64 +117,6 @@ class PipelineVerifiersTest(unittest.TestCase):
     self.assertTrue(mock_glob.called)
     self.assertEqual(verifiers.MAX_RETRIES + 1, mock_glob.call_count)
 
-  @patch.object(bigquery, 'Client')
-  def test_bigquery_matcher_success(self, mock_bigquery):
-    mock_query = Mock()
-    mock_client = mock_bigquery.return_value
-    mock_client.run_sync_query.return_value = mock_query
-    mock_query.fetch_data.return_value = ([], None, None)
-
-    matcher = verifiers.BigqueryMatcher(
-        'mock_project',
-        'mock_query',
-        'da39a3ee5e6b4b0d3255bfef95601890afd80709')
-    hc_assert_that(self._mock_result, matcher)
-
-  @patch.object(bigquery, 'Client')
-  def test_bigquery_matcher_query_run_error(self, mock_bigquery):
-    mock_query = Mock()
-    mock_client = mock_bigquery.return_value
-    mock_client.run_sync_query.return_value = mock_query
-    mock_query.run.side_effect = ValueError('job is already running')
-
-    matcher = verifiers.BigqueryMatcher('mock_project',
-                                        'mock_query',
-                                        'mock_checksum')
-    with self.assertRaises(ValueError):
-      hc_assert_that(self._mock_result, matcher)
-    self.assertTrue(mock_query.run.called)
-    self.assertEqual(verifiers.MAX_RETRIES + 1, mock_query.run.call_count)
-
-  @patch.object(bigquery, 'Client')
-  def test_bigquery_matcher_fetch_data_error(self, mock_bigquery):
-    mock_query = Mock()
-    mock_client = mock_bigquery.return_value
-    mock_client.run_sync_query.return_value = mock_query
-    mock_query.fetch_data.side_effect = ValueError('query job not executed')
-
-    matcher = verifiers.BigqueryMatcher('mock_project',
-                                        'mock_query',
-                                        'mock_checksum')
-    with self.assertRaises(ValueError):
-      hc_assert_that(self._mock_result, matcher)
-    self.assertTrue(mock_query.fetch_data.called)
-    self.assertEqual(verifiers.MAX_RETRIES + 1,
-                     mock_query.fetch_data.call_count)
-
-  @patch.object(bigquery, 'Client')
-  def test_bigquery_matcher_query_responds_error_code(self, mock_bigquery):
-    mock_query = Mock()
-    mock_client = mock_bigquery.return_value
-    mock_client.run_sync_query.return_value = mock_query
-    mock_query.run.side_effect = NotFound('table is not found')
-
-    matcher = verifiers.BigqueryMatcher('mock_project',
-                                        'mock_query',
-                                        'mock_checksum')
-    with self.assertRaises(NotFound):
-      hc_assert_that(self._mock_result, matcher)
-    self.assertTrue(mock_query.run.called)
-    self.assertEqual(verifiers.MAX_RETRIES + 1, mock_query.run.call_count)
 
 if __name__ == '__main__':
   logging.getLogger().setLevel(logging.INFO)
